Refactored activity layouts
