+++ app/src/main/AndroidManifest.xml
@@ -44,7 +44,8 @@
         <activity android:name=".PartnersActivity" />
         <activity android:name=".TourbookActivity"
             android:theme="@style/AppTheme.NoActionBar"/>
-        <activity android:name=".TourbookAscendActivity" />
+        <activity android:name=".TourbookAscendActivity"
+            android:theme="@style/AppTheme.NoActionBar"/>
         <activity android:name=".AboutActivity"
             android:theme="@style/AppTheme.NoActionBar"/>
     </application>
