+++ app/src/main/res/layout/activity_country.xml
@@ -16,8 +16,6 @@
 
         <include layout="@layout/appbar_empty"/>
 
-        <include layout="@layout/nav_tourbook"/>
-
         <include layout="@layout/content_table"/>
 
     </LinearLayout>
