+++ app/src/main/res/layout/appbar_empty.xml
@@ -2,10 +2,8 @@
 <android.support.design.widget.CoordinatorLayout
     xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".TourbookActivity">
+    android:layout_height="wrap_content">
 
     <android.support.design.widget.AppBarLayout
         android:layout_height="wrap_content"
@@ -21,5 +19,4 @@
 
     </android.support.design.widget.AppBarLayout>
 
-    <include layout="@layout/content_tourbook"/>
 </android.support.design.widget.CoordinatorLayout>
