+++ app/src/main/res/layout/appbar_info_and_map.xml
@@ -2,10 +2,8 @@
 <android.support.design.widget.CoordinatorLayout
     xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".TableActivity">
+    android:layout_height="wrap_content">
 
     <android.support.design.widget.AppBarLayout
         android:layout_height="wrap_content"
@@ -25,11 +23,18 @@
                 app:srcCompat="@android:drawable/ic_menu_info_details"
                 android:layout_gravity="end"
                 android:background="?attr/colorPrimary"
-                android:onClick="showComments"
-                />
+                android:onClick="showComments"/>
+
+            <ImageButton
+                android:layout_width="50dp"
+                android:layout_height="50dp"
+                app:srcCompat="@android:drawable/ic_menu_mapmode"
+                android:layout_gravity="end"
+                android:background="?attr/colorPrimary"
+                android:onClick="showMap"/>
+
         </android.support.v7.widget.Toolbar>
 
     </android.support.design.widget.AppBarLayout>
 
-    <include layout="@layout/content_table"/>
 </android.support.design.widget.CoordinatorLayout>
