+++ app/src/main/res/layout/comment_dialog.xml
@@ -4,8 +4,7 @@
     android:orientation="vertical"
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
-    android:layout_gravity="center"
-    android:id="@+id/dialogLayout">
+    android:layout_gravity="center">
 
     <RelativeLayout
         android:layout_width="match_parent"
@@ -29,7 +28,6 @@
     </RelativeLayout>
 
     <ScrollView xmlns:android="http://schemas.android.com/apk/res/android"
-        android:id="@+id/scrollView"
         android:layout_width="match_parent"
         android:layout_height="wrap_content">
