--- app/src/main/res/layout/content_tourbook.xml
@@ -1,101 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<android.support.constraint.ConstraintLayout android:id="@+id/baseLayout"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    app:layout_behavior="@string/appbar_scrolling_view_behavior"
-    android:background="#bbbbbb"
-    tools:showIn="@layout/app_bar_tourbook"
-    tools:context=".TourbookActivity">
-
-    <RelativeLayout
-        android:id="@+id/buttonLayout"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        app:layout_constraintTop_toTopOf="parent">
-
-        <ImageButton
-            android:id="@+id/prevYearButton"
-            android:layout_width="50dp"
-            android:layout_height="50dp"
-            android:layout_alignParentTop="true"
-            android:layout_alignParentLeft="true"
-            android:visibility="invisible"
-            android:onClick="goToPreviousYear"
-            app:srcCompat="@android:drawable/ic_media_rew" />
-
-        <TextView
-            android:id="@+id/yearTextView"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_centerInParent="true"
-            android:textStyle="bold"
-            android:textSize="20dp"/>
-
-        <ImageButton
-            android:id="@+id/nextYearButton"
-            android:layout_width="50dp"
-            android:layout_height="50dp"
-            android:layout_alignParentTop="true"
-            android:layout_alignParentRight="true"
-            android:visibility="invisible"
-            android:onClick="goToNextYear"
-            app:srcCompat="@android:drawable/ic_media_ff" />
-
-    </RelativeLayout>
-
-    <RadioGroup
-        android:id="@+id/radioButtonGroup"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:orientation="horizontal"
-        app:layout_constraintTop_toBottomOf="@+id/buttonLayout">
-
-        <RadioButton
-            android:id="@+id/ascendsRadioButton"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_weight="1"
-            android:buttonTint="@android:color/black"
-            android:checked="true"
-            android:onClick="showAscends"
-            android:text="Begehungen" />
-
-        <RadioButton
-            android:id="@+id/sackingsRadioButton"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_weight="1"
-            android:buttonTint="@android:color/black"
-            android:onClick="showBotches"
-            android:text="@string/botch_text_symbol"/>
-
-        <RadioButton
-            android:id="@+id/projectsRadioButton"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_weight="1"
-            android:buttonTint="@android:color/black"
-            android:onClick="showProjects"
-            android:text="@string/project_text_symbol"/>
-    </RadioGroup>
-
-    <ScrollView
-        android:id="@+id/scrollView"
-        android:layout_width="0dp"
-        android:layout_height="0dp"
-        app:layout_constraintTop_toBottomOf="@+id/radioButtonGroup"
-        app:layout_constraintLeft_toLeftOf="parent"
-        app:layout_constraintRight_toRightOf="parent"
-        app:layout_constraintBottom_toBottomOf="parent">
-
-        <LinearLayout
-            android:id="@+id/tableLayout"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:orientation="vertical">
-        </LinearLayout>
-    </ScrollView>
-</android.support.constraint.ConstraintLayout>
