+++ app/src/main/res/layout/export_dialog.xml
@@ -3,8 +3,7 @@
     android:orientation="vertical"
     android:layout_width="wrap_content"
     android:layout_height="wrap_content"
-    android:layout_gravity="center"
-    android:id="@+id/dialogLayout">
+    android:layout_gravity="center">
 
     <TextView
         android:layout_width="wrap_content"
@@ -39,8 +38,7 @@
         android:orientation="horizontal"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:layout_gravity="center"
-        android:id="@+id/externInternLayout">
+        android:layout_gravity="center">
 
         <Button xmlns:android="http://schemas.android.com/apk/res/android"
             android:layout_width="wrap_content"
