+++ app/src/main/res/layout/nav_header_main.xml
@@ -18,8 +18,7 @@
             android:layout_height="wrap_content"
             android:paddingTop="@dimen/nav_header_vertical_spacing"
             app:srcCompat="@mipmap/ic_launcher"
-            android:contentDescription="@string/nav_header_desc"
-            android:id="@+id/imageView"/>
+            android:contentDescription="@string/nav_header_desc"/>
 
     <TextView
             android:layout_width="match_parent"
@@ -31,7 +30,6 @@
     <TextView
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:text="@string/nav_header_subtitle"
-            android:id="@+id/textView"/>
+            android:text="@string/nav_header_subtitle"/>
 
 </LinearLayout>
