+++ app/src/main/res/menu/edit_delete.xml
@@ -1,0 +1,10 @@
+<?xml version="1.0" encoding="utf-8"?>
+<menu xmlns:android="http://schemas.android.com/apk/res/android">
+
+    <item
+        android:id="@+id/action_edit"
+        android:title="@string/edit" />
+    <item
+        android:id="@+id/action_delete"
+        android:title="@string/delete" />
+</menu>
