+++ app/src/main/res/menu/sync_delete.xml
